requirements.txt
requests==2.31.0
python-telegram-bot==20.7
schedule==1.2.0
python-dotenv==1.0.0
loguru==0.7.2
openai==1.12.0 